# api/openapi.yaml
openapi: 3.0.3
info:
  title: GoSheet API
  description: |
    HTTP API for GoSheet spreadsheet application.

    ## Authentication

    All `/api/*` endpoints require a bearer token. Two token types are in use:

    **Bootstrap token** (`BootstrapAuth`): Generated by the Go server at startup using a 32-byte
    CSPRNG (base64url encoded). Written to fd 3 alongside the port number. Electron injects it into
    the renderer, which sends it as `Authorization: Bearer <token>` on every request. Has full
    access to all endpoints. In `NODE_ENV=test` mode the server skips token generation and auth is
    disabled entirely so Playwright tests run without modification.

    **Agent token** (`AgentAuth`): Scoped token (`ro` or `rw`) issued by `POST /api/agent/token`
    in exchange for a valid bootstrap token. Used exclusively by AI agents. Agent tokens are
    rejected on `/api/file/*` endpoints regardless of scope. Read-only (`ro`) tokens are rejected
    on any mutating agent endpoint.

    The bootstrap endpoint (`GET /api/agent/bootstrap`) also accepts the agent token as a
    `?token=` query parameter for copy-paste UX when the user hands a URL to an agent.
  version: 1.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: Relative to host (localhost ephemeral port)

# All /api/* endpoints require auth. Static file routes are unauthenticated.
# In NODE_ENV=test the server disables auth entirely (bootstrapToken == "").
security:
  - BootstrapAuth: []
components:
  securitySchemes:
    BootstrapAuth:
      type: http
      scheme: bearer
      description: |
        Bootstrap token generated at server startup. Injected into the Electron renderer via URL
        query param and forwarded as `Authorization: Bearer <token>` on every API request.
        Has full access to all endpoints including `/api/file/*`.
    AgentAuth:
      type: http
      scheme: bearer
      description: |
        Scoped agent token issued by `POST /api/agent/token`. Used by AI agents.
        Scope `rw` allows reads and writes; scope `ro` allows reads only.
        Rejected on `/api/file/*` endpoints regardless of scope.
  schemas:
    ErrorResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code (e.g. INVALID_CELL_REF, FILE_NOT_FOUND)
...
    AgentScope:
      type: string
      enum: [ro, rw]
      description: |
        `ro` — read-only (GET agent endpoints only).
        `rw` — read + write (all agent endpoints).
    AgentTokenRequest:
      type: object
      required: [scope]
      properties:
        scope:
          $ref: '#/components/schemas/AgentScope'
    AgentTokenResponse:
      type: object
      required: [success, agentToken, agentId]
      properties:
        success:
          type: boolean
        agentToken:
          type: string
          description: Opaque bearer token for the agent to use on all subsequent calls.
        agentId:
          type: string
          description: Human-readable session identifier (e.g. `agt_a1b2c3d4`).
...
    AgentCellEntry:
      type: object
      required: [row, col, ref, raw, computed, isFormula, styleId, alignment]
      properties:
        row:
          type: integer
        col:
          type: integer
        ref:
          type: string
          description: A1 notation cell reference (e.g. `B3`).
        raw:
          type: string
          description: Raw value or formula string (e.g. `=A1+1` or `hello`).
        computed:
          type: string
          description: Evaluated string value.
        isFormula:
          type: boolean
        styleId:
          type: integer
        alignment:
          type: string
          enum: ['', left, center, right]
    AgentRangeResponse:
      type: object
      required: [success, range, cells]
      properties:
        success:
          type: boolean
        range:
          type: string
          description: The range that was queried (echoed back).
        cells:
          type: array
          items:
            $ref: '#/components/schemas/AgentCellEntry'
...
paths:
...
  /api/agent/token:
    post:
      operationId: issueAgentToken
      summary: Issue a scoped agent token
      description: |
        Exchanges the bootstrap token for a scoped agent token. Only one agent session
        may be active per file at a time — a second call returns 409 while a session is active.
      security:
        - BootstrapAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentTokenRequest'
      responses:
        '200':
          description: Agent token issued, session opened.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentTokenResponse'
        '400':
          description: Bad request (invalid scope)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (missing or invalid bootstrap token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict — an agent session is already active for this file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/agent/range:
    get:
      operationId: getAgentRange
      summary: Read cells in an A1-notation range
      description: |
        Returns raw value, computed value, formula flag, styleId, and alignment for every cell
        in the range. Empty cells are included with empty string values for predictable indexing.
        Maximum 10,000 cells per request.
      security:
        - AgentAuth: []
      parameters:
        - name: range
          in: query
          required: true
          description: A1 range notation, e.g. `A1:D10` or single cell `B3`.
          schema:
            type: string
      responses:
        '200':
          description: Cell data for the requested range.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRangeResponse'
        '400':
          description: Bad request (missing range param, invalid notation, range too large).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...